requirements.txt
future==1.0.0
iso8601==2.1.0
numpy==2.0.0
pyserial==3.5
PyYAML==6.0.1
pytest-8.3.4